Moved network selector from TopNavBar to separate view NetworkSelector.
